+++ s/index.ts
@@ -9,7 +9,8 @@ export * from "./common/views/id/view.js"
 export * from "./federated/utils/login.js"
 export * from "./federated/auth.js"
 export * from "./federated/views/components.js"
-export * from "./federated/views/auth-login/element.js"
+export * from "./federated/views/auth-user/element.js"
+export * from "./federated/views/auth-button/element.js"
 
 export * from "./tools/storage-signal.js"